Refactor: clean up __init__.py files for a sepration of application logic
